Move test_helper.py to tests/
