+++ tests/unit/api/entities/test_model.py
@@ -42,7 +42,7 @@ from otx.api.usecases.adapters.model_adapter import ModelAdapter
 from otx.api.utils.time_utils import now
 from tests.unit.api.constants.components import OtxSdkComponent
 from tests.unit.api.constants.requirements import Requirements
-from tests.unit.api.test_helpers import generate_random_single_image
+from tests.test_helpers import generate_random_single_image
 
 
 @pytest.mark.components(OtxSdkComponent.OTX_API)
